refactor: :truck: move code from internals to check
